Implementación de autenticacion finalizada con endpoints protegidos finalizada

+++ API_megaplay/Controllers/FavoritesController.cs
@@ -1,12 +1,17 @@
+using API_megaplay.Constants;
 using API_megaplay.Repository.IRepository;
 using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Cors;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 
 namespace API_megaplay.Controllers
 {
+    [Authorize]
     [Route("api/[controller]")]
     [ApiController]
+    [EnableCors(PolicyNames.AllowSpecificOrigin)] // Habilitar CORS para proteger este controlador
     public class FavoritesController : ControllerBase
     {
         private readonly IFavoritesRepository _favoritesRepository;
@@ -21,24 +26,39 @@ namespace API_megaplay.Controllers
 
 
         // Endpoint obtener las peliculas favoritas de un usuario mediante el ID
-        [HttpGet("user/{userId}/movies")]
-        public async Task<IActionResult> GetFavoriteMoviesByUser(int userId)
+        [HttpGet("movies")]
+        public async Task<IActionResult> GetFavoriteMoviesByUser()
         {
+            // Obtener el ID del usuario desde el token
+            var userIdClaim = User.FindFirst("id")?.Value;
+
+            if (string.IsNullOrEmpty(userIdClaim) || !int.TryParse(userIdClaim, out int userId))
+            {
+                return Unauthorized("No se pudo validar el usuario.");
+            }
+
             var favorites = await _favoritesRepository.GetFavoriteMoviesByUser(userId);
 
             if (favorites == null || !favorites.Any())
             {
-                return NotFound($"El usuario con ID {userId} no tiene películas favoritas.");
+                return NotFound($"El usuario no tiene películas favoritas.");
             }
 
             return Ok(favorites);
         }
 
+
         // Endpoint para agregar favoritos a un determinado usuario
         [HttpPost("add")]
-        public IActionResult AddToFavorites([FromQuery] int userId, [FromQuery] int movieId)
+        public IActionResult AddToFavorites([FromQuery] int movieId)
         {
-            // Validaciones personalizadas
+            var userIdClaim = User.FindFirst("id")?.Value;
+
+            if (string.IsNullOrEmpty(userIdClaim) || !int.TryParse(userIdClaim, out int userId))
+            {
+                return Unauthorized("No se pudo validar el usuario.");
+            }
+
             var userExists = _favoritesRepository.UserExists(userId);
             var movieExists = _favoritesRepository.MovieExists(movieId);
             var alreadyFavorite = _favoritesRepository.IsFavorite(userId, movieId);
@@ -60,10 +80,18 @@ namespace API_megaplay.Controllers
             return Ok("Película agregada a favoritos correctamente.");
         }
 
+
         // Endpoint para eliminar favoritos de un usuario
         [HttpDelete("remove")]
-        public IActionResult RemoveFromFavorites([FromQuery] int userId, [FromQuery] int movieId)
+        public IActionResult RemoveFromFavorites([FromQuery] int movieId)
         {
+            var userIdClaim = User.FindFirst("id")?.Value;
+
+            if (string.IsNullOrEmpty(userIdClaim) || !int.TryParse(userIdClaim, out int userId))
+            {
+                return Unauthorized("No se pudo validar el usuario.");
+            }
+
             var result = _favoritesRepository.RemoveFromFavorites(userId, movieId);
 
             if (!result)
@@ -73,7 +101,6 @@ namespace API_megaplay.Controllers
 
             return Ok("Película eliminada de favoritos correctamente.");
         }
-
 
     }
 }
